# Adding basic support for wildcards.