--- a/src/samples/hello_world/program.c
+++ b/src/samples/hello_world/program.c
@@ -18,6 +18,8 @@ void get_root(http_request* request, hw_http_response* response, void* user_data
     hw_string body;
     hw_string keep_alive_name;
     hw_string keep_alive_value;
+    hw_string route_matched_name;
+    hw_string route_matched_value;
     
     SETSTRING(status_code, HTTP_STATUS_200);
     hw_set_response_status_code(response, &status_code);
@@ -42,12 +44,18 @@ void get_root(http_request* request, hw_http_response* response, void* user_data
         hw_set_http_version(response, 1, 0);
     }
     
+    SETSTRING(route_matched_name, "Route-Matched");
+    route_matched_value.value = (char *) user_data;
+    route_matched_value.length = strlen((char *) user_data);
+    hw_set_response_header(response, &route_matched_name, &route_matched_value);
+    
     hw_http_response_send(response, "user_data", response_complete);
 }
 
 int main(int args, char** argsv)
 {
     char route[] = "/";
+    char sub_route[] = "/*";
     configuration config;
     config.http_listen_address = "0.0.0.0";
 
@@ -59,9 +67,11 @@ int main(int args, char** argsv)
     opt_flag_bool(conf, &config.tcp_nodelay, "tcp_nodelay", "If present, enables tcp_nodelay (i.e. disables Nagle's algorithm).");
     args = opt_config_parse(conf, args, argsv);
 
-    /* hw_init_from_config("hello_world.conf"); */
     hw_init_with_config(&config);
-    hw_http_add_route(route, get_root, NULL);
+    
+    hw_http_add_route(route, get_root, route);
+    hw_http_add_route(sub_route, get_root, sub_route);
+    
     hw_http_open();
 
     opt_config_free(conf);