This means that it will now be possible to map a handler to `/*` that will be used for requests such as `/foo` and `/bar/`. It will also allow you to define `/*/*` that will match `/foo/bar` but not `/foo`.