@kellabyte 2015-12-03
Awesome! I'll try to test this tomorrow and merge it :) I want to run some benchmarks against it.