@kellabyte 2015-12-07
I'm having some errors during benchmarks with a route of `/*/*`.

This seems to work.

```
curl http://104.130.19.205:8000/foo/bar
hello world
```

This seems to throw `Non-2xx or 3xx responses: 23108795`.

```
../bin/wrk/wrk --script ./pipelined_get.lua --latency -d 10s -t 40 -c 512 http://server:8000/foo/bar -- 32
Running 10s test @ http://server:8000/foo/bar
  40 threads and 512 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    17.40ms   12.36ms  83.69ms   54.93%
    Req/Sec    57.58k    14.63k  144.91k    69.70%
  Latency Distribution
     50%   16.79ms
     75%   28.14ms
     90%   34.95ms
     99%   39.75ms
  23108795 requests in 10.10s, 3.44GB read
  Non-2xx or 3xx responses: 2,310,8795
Requests/sec: 2,288,122.52
Transfer/sec:    349.11MB
```