@nmdguerreiro 2015-12-07
I'll have a look. I'm assuming you're getting a 404, but need to test it again.