@nmdguerreiro 2015-12-07
OK, so what's happening is that the lua script ignores the request path you're passing on the command line:

```lua
r[i] = wrk.format("GET", "/")
```

Which results in a `GET /`, as you can see from this tcpdump trace:

``` 
10:22:57.012231 IP (tos 0x0, ttl 64, id 61444, offset 0, flags [DF], proto TCP (6), length 1332, bad cksum 0 (->47bd)!)
    localhost.50769 > localhost.irdmi: Flags [P.], cksum 0x0329 (incorrect -> 0x5115), seq 25601:26881, ack 102401, win 9579, options [nop,nop,TS val 456822069 ecr 456822069], length 1280
E..4..@.@............Q.@f]z.\.J*..%k.).....
.:.5.:.5GET / HTTP/1.1
Host: localhost:8000
```