@kellabyte 2015-12-19
Can we rebase this to master? I think #78 to properly benchmark this at high throughput.