@kellabyte 2016-01-05
Ugh I'm really sorry but I refactored the route compare function. Could you rebase? I just made it a simple iteration because we aren't doing any complicated pattern matching yet with variables. Should be simple to check for your `*` character in the loop and return a match. I'm really sorry. Changing `url` to `hw_string` kind of messed up the old implementation and I realized we weren't really doing anything with tokenization.

Eventually this will be replicated by `pathfinder` with a trie structure but it's not ready yet.