@nmdguerreiro 2016-01-14
So, results (taken on two m4.xlarge with 4 threads for Haywire):

### master
```
[ec2-user@ip-172-31-5-244 wrk]$ ./wrk -c 300 -t 300 -d 10 -s pipelined_get.lua  --latency http://172.31.5.65:8000 -- 64
Running 10s test @ http://172.31.5.65:8000
  300 threads and 300 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    54.11ms  123.05ms   1.76s    90.47%
    Req/Sec     3.07k     1.64k   47.12k    78.98%
  Latency Distribution
     50%   12.36ms
     75%   26.77ms
     90%  171.76ms
     99%  647.16ms
  9454701 requests in 10.10s, 1.39GB read
  Socket errors: connect 0, read 0, write 0, timeout 4
Requests/sec: 936059.70
Transfer/sec:    141.05MB
```

### this branch
```
[ec2-user@ip-172-31-5-244 wrk]$ ./wrk -c 300 -t 300 -d 10 -s pipelined_get.lua  --latency http://172.31.5.65:8000 -- 64
Running 10s test @ http://172.31.5.65:8000
  300 threads and 300 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    55.25ms  122.46ms   1.75s    90.06%
    Req/Sec     3.11k     1.73k   58.17k    79.66%
  Latency Distribution
     50%   12.28ms
     75%   25.61ms
     90%  176.93ms
     99%  633.74ms
  9297647 requests in 10.10s, 1.37GB read
  Socket errors: connect 0, read 0, write 0, timeout 2
Requests/sec: 920558.02
Transfer/sec:    138.71MB
```

Note that the results for this branch do not account for the extra header I added to the example to determine which route is hit. I can remove that header if you think it's not useful. 

I think the relatively small degradation in performance is worth the added functionality.