@kellabyte 2016-01-14
I was curious why you didn't just check the last char for `== '*'` but I'm guessing you needed support for something like `/foo/*/bar/*`? Is that why you re-implemented the tokenization?