@nmdguerreiro 2016-01-14
Yeach, I just wanted to keep it flexible to allow stars at any point of the
url.
On Thu, 14 Jan 2016 at 20:10, Kelly Sommers <notifications@github.com>
wrote:

> Merged #67 <https://github.com/haywire/haywire/pull/67>.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/haywire/haywire/pull/67#event-515428686>.
>